third_party/blink/web_tests/crypto/subtle/hmac/sign-verify-expected.txt

Tests sign() and verify() for HMAC

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [fbdb1d1b18aa6c08324b7d64b71fb76370690e1d] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [fbdb1d1b18aa6c08324b7d64b71fb76370690e1d] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [fbdb1d1b18aa6c08324b7d64b71fb76370690e1d] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [fbdb1d1b18aa6c08324b7d64b71fb76370690e1d] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 8
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [b613679a0814d9ec772f95d778c35fc5ff1697c493715653c6c712144292c5ad] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 80
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [3c8162589aafaee024fc9a5ca50dd2336fe3eb28] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 80
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [3c8162589aafaee024fc9a5ca50dd2336fe3eb28] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 80
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [3c8162589aafaee024fc9a5ca50dd2336fe3eb28] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 80
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [3c8162589aafaee024fc9a5ca50dd2336fe3eb28] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 640
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4ac41ab89f625c60125ed65ffa958c6b490ea670] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 640
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4ac41ab89f625c60125ed65ffa958c6b490ea670] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 640
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4ac41ab89f625c60125ed65ffa958c6b490ea670] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-1"
PASS tmpKey.algorithm.length is 640
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4ac41ab89f625c60125ed65ffa958c6b490ea670] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 320
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [769f00d3e6a6cc1fb426a14a4f76c6462e6149726e0dee0ec0cf97a16605ac8b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 320
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [769f00d3e6a6cc1fb426a14a4f76c6462e6149726e0dee0ec0cf97a16605ac8b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 320
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [769f00d3e6a6cc1fb426a14a4f76c6462e6149726e0dee0ec0cf97a16605ac8b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 320
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [769f00d3e6a6cc1fb426a14a4f76c6462e6149726e0dee0ec0cf97a16605ac8b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 592
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4f1ee7cb36c58803a8721d4ac8c4cf8cae5d8832392eed2a96dc59694252801b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 592
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4f1ee7cb36c58803a8721d4ac8c4cf8cae5d8832392eed2a96dc59694252801b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 592
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4f1ee7cb36c58803a8721d4ac8c4cf8cae5d8832392eed2a96dc59694252801b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS tmpKey.type is "secret"
PASS typeof tmpKey.extractable is 'boolean'
PASS tmpKey.extractable is false
PASS tmpKey.algorithm.name is "HMAC"
PASS tmpKey.algorithm.hash.name is "SHA-256"
PASS tmpKey.algorithm.length is 592
PASS tmpKey.usages.join(',') is "sign,verify"
PASS: Mac should be [4f1ee7cb36c58803a8721d4ac8c4cf8cae5d8832392eed2a96dc59694252801b] and was
PASS verifyResult is true
PASS verifyResult is false
PASS successfullyParsed is true

TEST COMPLETE
